+++ arm_inlines_ascii.h
@@ -82,7 +82,7 @@ __INLINE CHAR toUpperCase(const CHAR ascii) {
 __INLINE CHAR INLINE__ HexToASCII(const BYTE hex)
 {
 	const BYTE h = hex & 0x0FU;
-	return (CHAR) (h) + ((h < 0x0AU) ? 0x30 : 0x37);
+	return (CHAR) (h + ((h < 0x0AU) ? 0x30U : 0x37U));
 }