+++ arm_inlines_average.h
@@ -20,7 +20,7 @@
 //! \a Justification: \p pArray is used as array, makes sense to declare it in the same manner.\n
 //! \a Justification: copy \c array needs to have a size and is always used in conjunction with \p nb.\n
 //! \b Rule-20.7 - \b Required: Enclosed macro parameters expansion (misra-c2012-20.7)\n
-//! \a Justification: \p typ_t parameter would be always be used as cast and wouldn't name concatenate.\n
+//! \a Justification: \p typ_t parameter would always be used as cast and wouldn't catenate name.\n
 // cppcheck-suppress-macro [misra-c2012-18.8, misra-c2012-20.7, misra-c2012-20.10]
 #define	RESTRICTED_AVERAGE(typ_t, sum_t)													\
 __INLINE typ_t RestrictedAverage_##typ_t(volatile const typ_t pArray[], const uint8_t nb)	\