+++ arm_inlines_ticks.h
@@ -18,12 +18,12 @@
 **			please refer to arm_chip_sam.h for an implementation example.
 ** \note	Define custom HAL_MS_TICKS_FACTOR at project level if tick period is not 1ms
 ** \param[in] prev - previously stored time value
-** \param[in] time - time lapse (in ms)
+** \param[in] lapse - time lapse (in ms)
 ** \return true if time elapsed
 **/
-__INLINE BOOL INLINE__ TPSSUP_MS(const DWORD prev, const DWORD time)
+__INLINE BOOL INLINE__ TPSSUP_MS(const DWORD prev, const DWORD lapse)
 {
-	const DWORD scaled_time = time * HAL_MS_TICKS_FACTOR;
+	const DWORD scaled_time = lapse * HAL_MS_TICKS_FACTOR;
 	const DWORD diff = OVF_DIFF(HALTicks(), prev);
 
 	return binEval(diff >= scaled_time);
@@ -35,13 +35,13 @@ __INLINE BOOL INLINE__ TPSSUP_MS(const DWORD prev, const DWORD time)
 **			please refer to arm_chip_sam.h for an implementation example.
 ** \note	Define custom HAL_MS_TICKS_FACTOR at project level if tick period is not 1ms
 ** \param[in] prev - previously stored time value
-** \param[in] time - time lapse (in ms)
+** \param[in] lapse - time lapse (in ms)
 ** \return true if time not elapsed
 **/
-__INLINE BOOL INLINE__ TPSINF_MS(const DWORD prev, const DWORD time)
+__INLINE BOOL INLINE__ TPSINF_MS(const DWORD prev, const DWORD lapse)
 {
 	const DWORD diff = OVF_DIFF(HALTicks(), prev);
-	const DWORD scaled_time = time * HAL_MS_TICKS_FACTOR;
+	const DWORD scaled_time = lapse * HAL_MS_TICKS_FACTOR;
 
 	return binEval(diff < scaled_time);
 }